+++ app/src/main/java/com/stopkaaaa/androidacademyproject/ui/MovieClickListener.kt
@@ -1,8 +1,6 @@
 package com.stopkaaaa.androidacademyproject.ui
 
-import com.stopkaaaa.androidacademyproject.data.models.Movie
-
 interface MovieClickListener {
-    fun movieClicked(movieId: Int)
+    fun movieClicked(movieId: Long)
     fun backPressed()
 }
